Feature/refactor search hook (#96)
* Refactor search hook in referral page

* Refactor referee search hook

* Remove useless import

* Refactor referee post

* Refactor referrer post search hook

* Remmove uesless import

* Change to error toast (#106)


---------
