+++ client/hooks/api/post/search-post.tsx
@@ -1,30 +1,69 @@
+import { ChangeEvent, useState } from "react"
 import apiService from "@/utils/common/api"
+import { postSortingOptions } from "@/utils/common/sorting/post"
 import { useInfiniteQuery } from "@tanstack/react-query"
 
 import { QueryKeyString } from "@/types/common/query-key-string"
 import { ReferralType } from "@/types/common/referral-type"
+import useDebounce from "@/hooks/common/debounce"
 
-interface ISearchPostFilterMeta {
-  companyName: string
-  cityUuid?: string
-  provinceUuid?: string
-  countryUuid?: string
-  industryUuid?: string
-  sorting: string
-  yoeMax?: string
-  yoeMin?: string
-}
-
-const useSearchPost = (
-  sorting: string,
-  filterMeta: ISearchPostFilterMeta,
-  type: ReferralType
-) => {
+const useSearchPost = (type: ReferralType) => {
   const keyString =
     type === ReferralType.REFEREE
       ? QueryKeyString.SEARCH_REFEREE_POST
       : QueryKeyString.SEARCH_REFERRER_POST
-  return useInfiniteQuery({
+
+  const [companyName, setCompanyName] = useState("")
+  const [provinceUuid, setProvinceUuid] = useState<undefined | string>()
+  const [countryUuid, setCountryUuid] = useState<undefined | string>()
+  const [cityUuid, setCityUuid] = useState<undefined | string>()
+  const [industryUuid, setIndustryUuid] = useState<undefined | string>()
+  const [yoeMin, setYoeMin] = useState<undefined | string>("0")
+  const [yoeMax, setYoeMax] = useState<undefined | string>("100")
+  const [sorting, setSorting] = useState(postSortingOptions[0].value)
+  const debouncedCompanyName = useDebounce(companyName, 800)
+
+  const handleCompanyChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setCompanyName(e.target.value)
+  }
+
+  const handleCountryChange = (value: string) => {
+    setCountryUuid(value)
+  }
+  const handleProvinceChange = (value: string) => {
+    setProvinceUuid(value)
+  }
+  const handleCityChange = (value: string) => {
+    setCityUuid(value)
+  }
+
+  const handleIndustryChange = (value: string) => {
+    setIndustryUuid(value)
+  }
+
+  const handleSortingChange = (value: string) => {
+    setSorting(value)
+  }
+
+  const handleYeoMinChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setYoeMin(e.target.value)
+  }
+
+  const handleYeoMaxChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setYoeMax(e.target.value)
+  }
+  const filterMeta = {
+    companyName: debouncedCompanyName,
+    cityUuid,
+    countryUuid,
+    industryUuid,
+    provinceUuid,
+    sorting,
+    yoeMin,
+    yoeMax,
+  }
+
+  const result = useInfiniteQuery({
     queryKey: [keyString, { sorting, filterMeta, type }],
     queryFn: apiService.searchPost,
     refetchOnWindowFocus: false,
@@ -37,6 +76,24 @@ const useSearchPost = (
       }
     },
   })
+  return {
+    result,
+    handleCompanyChange,
+    handleCountryChange,
+    handleProvinceChange,
+    handleCityChange,
+    handleSortingChange,
+    handleIndustryChange,
+    handleYeoMinChange,
+    handleYeoMaxChange,
+    provinceUuid,
+    cityUuid,
+    countryUuid,
+    industryUuid,
+    yoeMax,
+    yoeMin,
+    sorting,
+  }
 }
 
 export default useSearchPost
